+++ module_02/filter2d/workspace/Filter2D/src/host/filter2d.h
@@ -4,7 +4,7 @@
 #define FILTER2D_KERNEL_V_SIZE 	15
 
 void Filter2D(
-        const    int   coeffs[FILTER2D_KERNEL_V_SIZE][FILTER2D_KERNEL_H_SIZE],
+        const    short coeffs[FILTER2D_KERNEL_V_SIZE][FILTER2D_KERNEL_H_SIZE],
 		unsigned char *srcImg,
 		unsigned int   width,
 		unsigned int   height,
